\name{addtix}
\alias{addtix}

\title{add tix to plot}
\description{
 Add tick marks to edge of plot
}
\usage{
addtix(side = 3, pos = 0, tck = 0.005, at =c(0, 1), labels = FALSE, col = 2, ...)
}

\arguments{
  \item{side}{side of plot 1-4}
  \item{pos}{position relative to side}
  \item{tck}{tick size}
  \item{at}{locations along axis}
  \item{labels}{labels for tics}
  \item{col}{color for ticks}
  \item{\dots}{graphical parameters, par}
}

\value{
 Graphical Side Effect
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{par}

\keyword{aplot}
