\name{cleanWPX}
\alias{cleanWPX}
\title{Clean WPX
}
\description{
Return an empty (clean) WPX.

}
\usage{
cleanWPX()
}
\details{
Returns an empty list with NA's and 0's 
}
\value{
WPX list
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}
\note{Used internally.
}


\seealso{
addWPX,	catWPX,	checkWPX,repairWPX,clusterWPX,saveWPX, setWPX
}
\examples{

s0 = cleanWPX()

}
\keyword{misc}
