\name{COMPorder}
\alias{COMPorder}
\title{Seismic Component Order
}
\description{Set seismic component order
}
\usage{
COMPorder(STNS, COMPS)
}
\arguments{
  \item{STNS}{stations
}
  \item{COMPS}{components
}
}
\details{Sets up components so they are ordered
  according to V, N, E. used internally in swig.
}
\value{order vector
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\keyword{misc}
