\name{doMYBUTTS}
\alias{doMYBUTTS}

\title{Dummy Button Function}
\description{
 This is a dummy button function showing how buttons can be created on
 the fly
}
\usage{
doMYBUTTS(butt = "", clicks = NULL, x = NULL)
}

\arguments{
  \item{butt}{character vector}
  \item{clicks}{clicks}
  \item{x}{locations}
}

\keyword{misc}