\name{EmptyPickfile}
\alias{EmptyPickfile}
\title{Create an empty RSEIS pickfile structure}
\description{
Creates a structure list with no data
}
\usage{
EmptyPickfile(GH)
}
\arguments{
  \item{GH}{RSEIS list structure }
  }

\value{
 RSEIS pickfile list
}
\author{Jonathan M. Lees<jonathan.lees.edu>}
\seealso{EmptySEIS}
\examples{

data(GH)
EmptyPickfile(GH)


}

\keyword{misc}
