\name{fixUWstasLL}
\alias{fixUWstasLL}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{fixUWstasLL}
\description{
  Matches station locations  to pickfile stations
}
\usage{
fixUWstasLL(STAS, stafile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{STAS}{structure of station lat, lon, z}
  \item{stafile}{station file}
}
\details{
Matches station locations  to pickfile stations
}
\value{
 structure of station lat, lon, z
}
\author{Jonathan M. Lees<jonathan.lees.edu>}

\keyword{misc}
