\name{getFcard}
\alias{getFcard}
\title{Parse UW F Card}
\description{
  get F-card information
}
\usage{
getFcard(FCARD)
}
\arguments{

  \item{FCARD}{Error Ellipsoid card }
}

\value{

  List:
      \item{azim1}{angle, degrees }
      \item{plunge1}{angle, degrees }
      \item{val1}{value }
      
      \item{azim2}{angle, degrees }
      \item{plunge2}{angle, degrees }
      \item{val2}{value }
   
      
      \item{azim3}{angle, degrees }
      \item{plunge3}{angle, degrees }
      \item{val3}{value }
      \item{herr}{error }
     \item{ verr}{vertical error }
 
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\keyword{misc}
