\name{getjul}
\alias{getjul}

\title{Get Julian day}
\description{
  Get Julian day
}
\usage{
getjul(year, month, day)
}

\arguments{
  \item{year}{year}
  \item{month}{month}
  \item{day}{day of month }
}

\value{
Julian Day
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{getmoday}
\examples{
getjul(2003, 11, 13)
}

\keyword{misc}
