\name{getmoday}
\alias{getmoday}

\title{Get Month Day}
\description{
 Get month day from julian day and year
}
\usage{
getmoday(jul, iyear)
}

\arguments{
  \item{jul}{julian day}
  \item{iyear}{Year}
}

\value{
  \item{mo}{Month}
   \item{ dom}{day of month}
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\examples{
getmoday(234, 2005)
}

\keyword{misc}
