\name{getrdpix}
\alias{getrdpix}
\title{get read picks
}
\description{get read picks
}
\usage{
getrdpix(zloc, zenclick, sel, NH)
}
\arguments{
  \item{zloc}{location list
}
  \item{zenclick}{number of picks
}
  \item{sel}{sel vector in swig
}
  \item{NH}{RSEIS list
}
}
\details{Used internally in swig
}
\value{list: rd: date/times of picks for stations and comps
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\seealso{swig
}

\keyword{misc}
