\name{GLUE.GET.seis}
\alias{GLUE.GET.seis}

\title{GLUE.GET.seis}
\description{
  Once a database has been mined this program re-arranges the
  seismograms and creates a structure used in other programs.
}
\usage{
GLUE.GET.seis(GG)
}

\arguments{
  \item{GG}{list of seismograms with headers}
}

\value{
 structure of seismograms glued together
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{Mine.seis}


\keyword{misc}
