\name{hypot}
\alias{hypot}
\title{Hypot}

\description{
 length of line connecting two points in a plane
}
\usage{
hypot(x1, y1, x2, y2)
}

\arguments{
  \item{x1}{x-location point 1 }
  \item{y1}{ y-location point 1 }
  \item{x2}{x-location point 2  }
  \item{y2}{ y-location point 2 }
}
\details{
  Euclidean distance
}
\value{
 numeric distance
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\examples{

hypot(34, 12, 56, 89)

}

\keyword{misc }
