\name{info.seis}
\alias{info.seis}
\title{Information on a Seismic record
}
\description{Retrieve information on a seismic record
}
\usage{
info.seis(GH)
}
\arguments{
  \item{GH}{RSEIS seismic record list
}
}
\details{Prints summary infromation on the traces in the seismic record
}
\value{Side Effects
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\examples{

data(KH)
info.seis(KH)


}
\keyword{misc}
