\name{jpolyval}
\alias{jpolyval}

\title{Polynomial Value}
\description{
 Polynomila value
}
\usage{
jpolyval(p, x)
}

\arguments{
  \item{p}{coefficients}
  \item{x}{input value}
}

\value{
 Sum of polynomial: \eqn{p_1+p_2*x^1+p_3*x^2...}
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\examples{
jpolyval(c(2,3,5), 7)

}

\keyword{misc}
