\name{legitpix}
\alias{legitpix}
\title{Legitimate picks in swig
}
\description{Legitimate picks in swig (used internally)
}
\usage{
legitpix(sel, zloc, zenclick)
}
\arguments{
  \item{sel}{seleceted traces in swig
}
  \item{zloc}{location list
}
  \item{zenclick}{number of legitimate picks
}
}

\value{list: ypick, ppick
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\seealso{swig
}

\keyword{misc}
