\name{NEW.getUWSTAS}
\alias{NEW.getUWSTAS}

\title{get UW station file}
\description{
  Match Picks with stations and return station structure
}
\usage{
NEW.getUWSTAS(PICS)
}

\arguments{
  \item{PICS}{Picks in pickfile}
}
\details{
  matches Picks with stations
}
\value{
  STAS structure
}

\author{Jonathan M. Lees<jonathan.lees.edu>}


\keyword{misc }