\name{PDE2list}
\Rdversion{1.1}
\alias{PDE2list}
\title{Convert PDEs to List
}
\description{Converta list of individual PDE events
  to a list of lat, lon, z...etc
}
\usage{
PDE2list(PDF)
}
\arguments{
  \item{PDF}{list of individual events
}
}
\details{uses getmem
}
\value{list
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{getmem, getPDEcsv, parse.pde,getPDEscreen
}

\keyword{misc}
