\name{PPIX}
\alias{PPIX}

\title{P-picking}
\description{
 Add Pick Marks and Labels
}
\usage{
PPIX(zloc, YN = NULL, col = 1, lab = "")
}

\arguments{
  \item{zloc}{locator output}
  \item{YN}{number of panels}
  \item{col}{color for picks}
  \item{lab}{labels for picks}
}

\value{
  Side Effects
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\keyword{aplot}
