\name{Put1Dvel}
\alias{Put1Dvel}

\title{Dump a velocity model to an ascii file}
\description{
 Dump a velocity model to an ascii file
}
\usage{
Put1Dvel(vel, outfile)
}

\arguments{
  \item{vel}{Velocity Model Structure}
  \item{outfile}{File name}
}

\value{
Side Effects
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\seealso{Get1Dvel, travel.time1D}

\keyword{misc}
