\name{readUW.OSTAS}
\alias{readUW.OSTAS}

\title{Parse UW  O-Cards}
\description{
  PArse out UW O-cards from Pickfile
}
\usage{
readUW.OSTAS(OS1)
}

\arguments{
  \item{OS1}{cards starting with O}
}

\value{
  vector of station names not picked
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\keyword{misc}