\name{RSEIS-package}
\alias{RSEIS-package}
\alias{RSEIS}
\docType{package}
\title{
Seismic Analysis and Display
}
\description{
  Multiple interactive codes to analize seismic data, via, spectrum
  analysis, wavelet transforms, particle motion, hodograms.
}
\details{
\tabular{ll}{
Package: \tab RSEIS\cr
Type: \tab Package\cr
Version: \tab 1.0-9\cr
Date: \tab 2008-04-15\cr
License: \tab GPL\cr
}

PICK.GEN is general all pupose code that uses interaction for seismic analysis.
}
\author{
Jonathan M. Lees<jonathan.lees.edu>
Maintainer:Jonathan M. Lees<jonathan.lees.edu>
}
\references{

}

\keyword{ package }
\seealso{
RPGM, RFOC
}
\examples{
data("GH")
PICK.GEN(GH)

}

