\name{screens}
\alias{screens}

\title{screens}
\description{
 Open n devices for plotting.
}
\usage{
screens(n)
}

\arguments{
  \item{n}{number of devices required}
}
\details{
  If k screens are open and k >= n, nothing is done.
}
\value{
 Graphical Side Effect
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{devices}
\examples{
screens(2)
}

\keyword{misc}
