\name{selstas}
\alias{selstas}
\title{Select Stations
}
\description{Extract a set of stations from a
  longer station file.
}
\usage{
selstas(sta, ind)
}
\arguments{
  \item{sta}{station list (name, lat, lon, z)
}
  \item{ind}{index to station list = positive is select, negative is remove
}
}

\value{ station list with those indeces either removed or save.
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\keyword{misc}
