\name{T12.pix}
\alias{T12.pix}

\title{Get T1, T2}
\description{
Modify opick data frame and add T2=T1+dur
}
\usage{
T12.pix(A)
}

\arguments{
  \item{A}{pick data.frame }
}
\details{
 Given t1 and duration, returns to structure, t2=t1+dur.
}
\value{
  pick data.frame with t2 as a member.
  
}
\author{Jonathan M. Lees<jonathan.lees.edu>}


\keyword{misc}
