\name{tojul}
\alias{tojul}

\title{Julian Day}
\description{
 Convert to Julian Day. Used for calculations.
}
\usage{
tojul(year, month, day)
}

\arguments{
  \item{year}{year}
  \item{month}{month}
  \item{day}{day}
}

\value{
 Julian Days
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\examples{
tojul(1953, 3, 19)
}

\keyword{misc}
