\name{tomo.colors}
\alias{tomo.colors}

\title{Tomography Colors}
\description{
 Color Palette ranging from red to blue through black.
}
\usage{
tomo.colors(n, alpha = 1)
}

\arguments{
  \item{n}{number of colors }
  \item{alpha}{hsv color parameter}
}
\details{

}
\value{
color palette
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{rainbow, colors, hsv}
\examples{
tomo.colors(25, alpha = 1)
}

\keyword{misc}
