\name{writeUW.Commentcard}
\alias{writeUW.Commentcard}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{writeUW.Commentcard}
\description{
  write UW pickfile
}
\usage{
writeUW.Commentcard(comments)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{comments}{comment vector }
}
\value{
 Side Effects
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\keyword{misc}
