\name{writeUW.DOTcard}
\alias{writeUW.DOTcard}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{writeUW.DOTcard}
\description{
  write UW pickfile
}
\usage{
writeUW.DOTcard(STAS)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{STAS}{station structure}
}
\value{
 Side Effects
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\keyword{misc}
