\name{writeUW.Fcard}
\alias{writeUW.Fcard}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{writeUW.Fcard}
\description{
  write UW pickfile
}
\usage{
writeUW.Fcard(F)
}

\arguments{
  \item{F}{F-card info}
}
\value{
 Side Effects
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\keyword{misc}