\name{writeUWpickfile}
\alias{writeUWpickfile}

\title{UW formatted ascii pickfile}
\description{
Write UW formatted ascii pickfile
}
\usage{
writeUWpickfile(A, output = "")
}

\arguments{
  \item{A}{Pickfile structure}
  \item{output}{output file}
}

\value{
 Side Effects. Used to save ASCII versions of pickfiles for other processing.
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{EmptyPickfile}

\keyword{misc}
