\name{DO.PMOT.ARR}
\alias{DO.PMOT.ARR}

\title{Particle Motion Analysis with arrows}
\description{
 Plot particle motion arrows
}
\usage{
DO.PMOT.ARR(E, N)
}

\arguments{
  \item{E}{East component}
  \item{N}{East Component}
}

\value{
Graphical Side Effects
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{PMOT.drive}
\examples{


\dontrun{ DO.PMOT.ARR(x,y) }

}

\keyword{misc}
\keyword{aplot}
