\name{fixUWstasLL}
\alias{fixUWstasLL}

\title{fixUWstasLL}
\description{
  Matches station locations  to pickfile stations
}
\usage{
fixUWstasLL(STAS, stafile)
}

\arguments{
  \item{STAS}{structure of station lat, lon, z}
  \item{stafile}{station file}
}
\details{
Matches station locations  to pickfile stations
}
\value{
 structure of station lat, lon, z
}
\author{Jonathan M. Lees<jonathan.lees.edu>}

\keyword{misc}
