\name{hypot}
\alias{hypot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Hypoteneuse }
\description{
 length of line connecting two points in a plane
}
\usage{
hypot(x1, y1, x2, y2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x1}{x-location point 1 }
  \item{y1}{ y-location point 1 }
  \item{x2}{x-location point 2  }
  \item{y2}{ y-location point 2 }
}
\details{
  Euclidean distance
}
\value{
 numeric distance
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\examples{

hypot(34, 12, 56, 89)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc }
