\name{PICK.DOC}
\alias{PICK.DOC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Documentation for PICK.GEN}
\description{
Prints brief documentation for buttons in PICK.GEN
}
\usage{
PICK.DOC(w)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{w}{vector of buttons needed}
}
\details{
Buttons are defined in advance
}
\value{
printed side effect
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{PICK.GEN}
\examples{
PICK.DOC(6:23)
}

\keyword{misc}
