\name{PICK.DOC}
\alias{PICK.DOC}

\title{Documentation for swig}
\description{
Prints brief documentation for buttons in swig
}
\usage{
PICK.DOC(w)
}

\arguments{
  \item{w}{vector of buttons needed}
}
\details{
Buttons are defined in advance
}
\value{
printed side effect
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{swig}
\examples{
PICK.DOC(6:23)
}

\keyword{misc}
